🔀 Implement v1 of the API. (#2)
V1 of the API supports the same requests as the test version, but now with users! Todos are now aware of what userid created them, and are now accessed and updated with an extra url param: `api/v1/todos/`**`:user`**`/:id`

Both `andy` and `trymunx` are hardcoded as users in the database, but hitting `api/v1/user` will generate a new userid for the client to use.